+++ ale/drivers/mro_drivers.py
@@ -1,3 +1,4 @@
+import numpy as np
 import spiceypy as spice
 
 from ale.base import Driver
@@ -180,6 +181,12 @@ class MroMarciIsisLabelNaifSpiceDriver(LineScanner, IsisLabel, NaifSpice, NoDist
                 line = 0.5
             self._ephemeris_stop_time = max(self.compute_marci_time(line))
         return self._ephemeris_stop_time
+    
+    @property
+    def ephemeris_time(self):
+        if not hasattr(self, "_ephemeris_time"):
+            self._ephemeris_time = np.linspace(self.ephemeris_start_time, self.ephemeris_stop_time + self.interframe_delay, self.image_lines + 1)
+        return self._ephemeris_time
 
     @property
     def detector_center_line(self):